requirements.txt
# daily-intel dependencies
# Install: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0

# LLM
anthropic>=0.40.0
google-generativeai>=0.8.0   # Gemini fallback

# Financial data
yfinance>=0.2.40

# Geopolitical / RSS
feedparser>=6.0.11

# Social
atproto>=0.0.54               # Bluesky

# Dark web proxy
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Email templates
jinja2>=3.1.0

# Scheduler
APScheduler>=3.10.4
pytz>=2024.1
